chore(pkg/scripts_test): clean up
