+++ pkg/scripts_test/consts_linux.go
@@ -1,0 +1,10 @@
+package sumologic_scripts_tests
+
+const (
+	rootGroup            string = "root"
+	systemdDirectoryPath string = "/run/systemd/system"
+	systemdPath          string = "/etc/systemd/system/otelcol-sumo.service"
+
+	systemUser  string = "otelcol-sumo"
+	systemGroup string = "otelcol-sumo"
+)
