+++ pkg/scripts_test/install_darwin_test.go
@@ -76,7 +76,7 @@ func TestInstallScriptDarwin(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions(getSystemUser(), getSystemGroup()),
+				checkConfigFilesOwnershipAndPermissions(systemUser, systemGroup),
 				checkUserConfigCreated,
 				checkLaunchdConfigCreated,
 			},
@@ -91,11 +91,12 @@ func TestInstallScriptDarwin(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions(getSystemUser(), getSystemGroup()),
+				checkConfigFilesOwnershipAndPermissions(systemUser, systemGroup),
 				checkUserConfigCreated,
 				checkLaunchdConfigCreated,
 				checkTokenInLaunchdConfig,
 				checkUserExists,
+				checkGroupExists,
 				checkHostmetricsConfigNotCreated,
 			},
 		},
@@ -110,13 +111,14 @@ func TestInstallScriptDarwin(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions(getSystemUser(), getSystemGroup()),
+				checkConfigFilesOwnershipAndPermissions(systemUser, systemGroup),
 				checkUserConfigCreated,
 				checkLaunchdConfigCreated,
 				checkTokenInLaunchdConfig,
 				checkUserExists,
+				checkGroupExists,
 				checkHostmetricsConfigCreated,
-				checkConfigFilesOwnershipAndPermissions(getSystemUser(), getSystemGroup()),
+				checkConfigFilesOwnershipAndPermissions(systemUser, systemGroup),
 			},
 		},
 		{
@@ -132,11 +134,12 @@ func TestInstallScriptDarwin(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions(getSystemUser(), getSystemGroup()),
+				checkConfigFilesOwnershipAndPermissions(systemUser, systemGroup),
 				checkUserConfigCreated,
 				checkLaunchdConfigCreated,
 				checkTokenInLaunchdConfig,
 				checkUserExists,
+				checkGroupExists,
 			},
 		},
 		{
@@ -145,17 +148,30 @@ func TestInstallScriptDarwin(t *testing.T) {
 				installToken: installToken,
 			},
 			preActions: []checkFunc{preActionMockLaunchdConfig},
-			preChecks:  []checkFunc{checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated, checkUserNotExists, checkGroupNotExists, checkLaunchdConfigCreated},
-			postChecks: []checkFunc{checkBinaryCreated, checkBinaryIsRunning, checkConfigCreated, checkUserConfigCreated, checkLaunchdConfigCreated, checkTokenInLaunchdConfig},
+			preChecks: []checkFunc{
+				checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated,
+				checkUserNotExists, checkGroupNotExists, checkLaunchdConfigCreated,
+			},
+			postChecks: []checkFunc{
+				checkBinaryCreated, checkBinaryIsRunning, checkConfigCreated,
+				checkUserConfigCreated, checkLaunchdConfigCreated, checkTokenInLaunchdConfig,
+			},
 		},
 		{
 			name: "different installation token",
 			options: installOptions{
 				installToken: installToken,
 			},
-			preActions:  []checkFunc{preActionMockLaunchdConfig, preActionWriteDifferentTokenToLaunchdConfig},
-			preChecks:   []checkFunc{checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated, checkUserNotExists, checkGroupNotExists, checkLaunchdConfigCreated},
-			postChecks:  []checkFunc{checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated, checkUserNotExists, checkGroupNotExists, checkLaunchdConfigCreated, checkAbortedDueToDifferentToken},
+			preActions: []checkFunc{preActionMockLaunchdConfig, preActionWriteDifferentTokenToLaunchdConfig},
+			preChecks: []checkFunc{
+				checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated,
+				checkUserNotExists, checkGroupNotExists, checkLaunchdConfigCreated,
+			},
+			postChecks: []checkFunc{
+				checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated,
+				checkUserNotExists, checkGroupNotExists, checkLaunchdConfigCreated,
+				checkAbortedDueToDifferentToken,
+			},
 			installCode: 1,
 		},
 	} {
