+++ pkg/scripts_test/install_unix_test.go
@@ -56,7 +56,7 @@ func TestInstallScript(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions("root", getRootGroupName()),
+				checkConfigFilesOwnershipAndPermissions("root", rootGroup),
 				checkUserConfigNotCreated,
 				checkSystemdConfigNotCreated,
 			},
@@ -83,7 +83,7 @@ func TestInstallScript(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions("root", getRootGroupName()),
+				checkConfigFilesOwnershipAndPermissions("root", rootGroup),
 				checkUserConfigCreated,
 				checkTokenInConfig,
 				checkSystemdConfigNotCreated,
@@ -103,7 +103,7 @@ func TestInstallScript(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions("root", getRootGroupName()),
+				checkConfigFilesOwnershipAndPermissions("root", rootGroup),
 				checkUserConfigCreated,
 				checkDeprecatedTokenInConfig,
 				checkSystemdConfigNotCreated,
@@ -123,13 +123,13 @@ func TestInstallScript(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions("root", getRootGroupName()),
+				checkConfigFilesOwnershipAndPermissions("root", rootGroup),
 				checkUserConfigCreated,
 				checkTokenInConfig,
 				checkSystemdConfigNotCreated,
 				checkUserNotExists,
 				checkHostmetricsConfigCreated,
-				checkHostmetricsOwnershipAndPermissions("root", getRootGroupName()),
+				checkHostmetricsOwnershipAndPermissions("root", rootGroup),
 			},
 		},
 		{
@@ -146,7 +146,7 @@ func TestInstallScript(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions("root", getRootGroupName()),
+				checkConfigFilesOwnershipAndPermissions("root", rootGroup),
 				checkUserConfigCreated,
 				checkTokenInConfig,
 				checkSystemdConfigNotCreated,
@@ -266,7 +266,7 @@ func TestInstallScript(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
-				checkConfigFilesOwnershipAndPermissions("root", getRootGroupName()),
+				checkConfigFilesOwnershipAndPermissions("root", rootGroup),
 				checkTags,
 				checkSystemdConfigNotCreated,
 			},
